+++ tests/data/ex_no_split_subscripts.py
@@ -1,0 +1,100 @@
+x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+
+x["xxxxxxxxxxxxxxxxxx"]["xxxxxxxxxxxxxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+
+x[x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] or "xxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxx"
+
+x["xxxxxxxxxx", "xxxxxxxxxx", "xxxxxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+
+x[x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] or "xxxxxxx"] = "x"
+
+x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = "x"
+
+x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = "x"
+
+xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx = xxxxxxxxxx[xxxxxxxxxx]
+
+xxxxxxxxxxx = xxxxxxxxxxxxxxx[x(xxxxxxxxxxx[xxxxxxxxxxx].xxxxxxxxxxxx(xxxxxxxxxxxxxxxxxxxxxxxxx))]
+
+xxxxxxxxxxx = xxxxxxxxxxxxxxx[x(xxxxxxxxxxx[xxxxxxxxxxx].xxxxxxxxxxxx(xxxxxxxxxxxxxxxxxxxxxxxxxx))]
+
+xxxxxxxxxx["xxxxxxxxxx"] = xxxxxxxxxx["xxxxxxxxxx"] = xxxxxxxxxx["xxxxxxxxxx"] = xxxxxxxxxx["xxxxxxxxxx"] = xxxxxxxxxx["xxxxxxxxxx"] = xxxxxxxxxx["xxxxxxxxxx"]
+
+x[
+    0
+    # comment
+] = x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+
+if xxxxxxxxxxxxxxxxxxxxxxxxxx in ["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"]:
+    pass
+
+return xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx[xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx]
+
+return xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx[xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx]
+
+# output
+
+x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = (
+    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+)
+
+x["xxxxxxxxxxxxxxxxxx"]["xxxxxxxxxxxxxxxxxx"] = (
+    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+)
+
+x[x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] or "xxxxxxx"] = (
+    "xxxxxxxxxxxxxxxxxxxxxxxxx"
+)
+
+x["xxxxxxxxxx", "xxxxxxxxxx", "xxxxxxxxxx"] = (
+    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+)
+
+x[
+    x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"]
+    or "xxxxxxx"
+] = "x"
+
+x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = (
+    "x"
+)
+
+x[
+    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+] = "x"
+
+xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx = (
+    xxxxxxxxxx[xxxxxxxxxx]
+)
+
+xxxxxxxxxxx = (
+    xxxxxxxxxxxxxxx[x(xxxxxxxxxxx[xxxxxxxxxxx].xxxxxxxxxxxx(xxxxxxxxxxxxxxxxxxxxxxxxx))]
+)
+
+xxxxxxxxxxx = xxxxxxxxxxxxxxx[
+    x(xxxxxxxxxxx[xxxxxxxxxxx].xxxxxxxxxxxx(xxxxxxxxxxxxxxxxxxxxxxxxxx))
+]
+
+xxxxxxxxxx["xxxxxxxxxx"] = (
+    xxxxxxxxxx["xxxxxxxxxx"]
+) = (
+    xxxxxxxxxx["xxxxxxxxxx"]
+) = xxxxxxxxxx["xxxxxxxxxx"] = xxxxxxxxxx["xxxxxxxxxx"] = xxxxxxxxxx["xxxxxxxxxx"]
+
+x[
+    0
+    # comment
+] = (
+    x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"]
+) = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+
+if xxxxxxxxxxxxxxxxxxxxxxxxxx in [
+    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+]:
+    pass
+
+return xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx[xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx]
+
+return xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx[
+    xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+]