+++ tests/test_format_ex.py
@@ -28,6 +28,10 @@ class TestSimpleFormatEX(BlackBaseTestCase):
             black.Mode(keep_blank_lines_in_brackets=True),
         )
 
+    @patch("black.dump_to_file", dump_to_stderr)
+    def test_no_split_subscripts(self) -> None:
+        self.check_file("ex_no_split_subscripts", black.Mode(split_subscripts=False))
+
     def check_file(self, filename: str, mode: black.Mode, *, data: bool = True) -> None:
         source, expected = read_data(filename, data=data)
         actual = fs(source, mode=mode)